Hack basic example to work
Get it to at least load a shader that can set the frag color.

Anything with images will fail until the regexp parser is fixed.

+++ src/ofxISF/Constants.h
@@ -20,10 +20,11 @@ struct Ref_ : public ofPtr<T>
 template <typename T>
 struct Type2Int
 {
-	static unsigned int value()
+	static size_t value()
 	{
 		static size_t m = 0;
-		return (unsigned int)&m;
+        //return dynamic_cast<unsigned int>(&m);
+		return (size_t)&m;
 	}
 };
